requirements.txt
pillow==9.1.0
matplotlib
mysql-connector-python
streamlit
